+++ packages/rainbowkit/src/types/utils.ts
@@ -60,7 +60,8 @@ export type WalletProviderFlags =
   | 'isCoin98'
   | 'isMEWwallet'
   | 'isSafeheron'
-  | 'isSafePal';
+  | 'isSafePal'
+  | '__seif';
 
 export type WalletProvider = Evaluate<
   EIP1193Provider & {