+++ packages/rainbowkit/src/wallets/walletConnectors/magicEdenWallet/magicEdenWallet.ts
@@ -8,6 +8,7 @@ export const magicEdenWallet = (): Wallet => {
   return {
     id: 'magicEden',
     name: 'Magic Eden Wallet',
+    rdns: 'io.magiceden.wallet',
     iconUrl: async () => (await import('./magicEden.svg')).default,
     iconBackground: '#36114D',
     installed: hasInjectedProvider({ namespace: 'magicEden.ethereum' }),