+++ packages/rainbowkit/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rainbow-me/rainbowkit
 
+## 2.1.3
+
+### Patch Changes
+
+- 23e33b9: Added `mantle` and `mantleTestnet` network support
+- 7b00be5: Added Seif Wallet support with `seifWallet` wallet connector
+- 001a0a9: Resolved an issue in development where browser detection would throw an error if `navigator.userAgent` was unavailable in the browser.
+
 ## 2.1.2
 
 ### Patch Changes